Tracking:  - Implement `LocationService`.  - Implement the `LocationRepository` to contain the current user location and tracking state.  - Implement the Tracking screen UI and ViewModel.  - Connect it with the repositories and `LocationService`.  - Minor refactoring in the DB.  - Minor refactoring the repositories.

+++ core-data/src/main/java/android/template/core/data/TripModel.kt
@@ -9,5 +9,5 @@ data class TripModel(
     val id: Long,
     val startDate: Date,
     val duration: Long,
-    val distance: Double
+    val distance: Int
 )
